@@ -68,7 +68,7 @@
         </div>
     </div>
 
-    <table class="table w-100 mt">
+    <table class="table w-100 mt persons">
         <thead>
             <tr>
                 <th>
