@@ -3,14 +3,14 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Entities;
 
-public class PersonsDbContext : DbContext
+public class ApplicationDbContext : DbContext
 {
-    public PersonsDbContext(DbContextOptions options)
+    public ApplicationDbContext(DbContextOptions options)
         : base(options) { }
 
-    public DbSet<Country> Countries { get; set; }
+    public virtual DbSet<Country> Countries { get; set; }
 
-    public DbSet<Person> Persons { get; set; }
+    public virtual DbSet<Person> Persons { get; set; }
 
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
