@@ -1,4 +1,4 @@
-﻿// <auto-generated />
+// <auto-generated />
 using System;
 using Entities;
 using Microsoft.EntityFrameworkCore;
@@ -10,7 +10,7 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 
 namespace Entities.Migrations
 {
-    [DbContext(typeof(PersonsDbContext))]
+    [DbContext(typeof(ApplicationDbContext))]
     [Migration("20250219204037_TINColumn")]
     partial class TINColumn
     {
