@@ -28,5 +28,5 @@ public class Person
 
     public string? TIN { get; set; }
 
-    public Country? Country { get; set; }
+    public virtual Country? Country { get; set; }
 }
